Added NO_LABEL constant to help with consistency, moved variables to common.constants
